requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pandas==2.3.2
openpyxl==3.1.5
streamlit==1.36.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
cloudscraper==1.2.71
# Optional: API classification (uncomment when needed)
# openai>=1.0.0
# anthropic>=0.7.0
